requirements.txt
Flask>=2.0
python-dotenv
azure-cosmos
azure-storage-blob
